pyproject.toml
[tool.poetry]
name = "sentinel-airflow-spark"
version = "0.1.0"
description = ""
authors = ["thattommyhall <thattommyhall@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pendulum = "^2.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
